Version 0.7

+++ src/control/include/control.h
@@ -3,6 +3,6 @@
 
 bool is_pressed_left(int* mouse_x, int* mouse_y);
 bool is_pressed_enter();
-void change_scale(float* scale);
+void change_scale(double* scale);
 
 #endif
