+++ src/parser/include/parsing.h
@@ -8,10 +8,10 @@
 void parse_input(std::string* input_x, std::string* input_y);
 
 struct equations {
-    inline static float x;
-    inline static float y;
-    inline static exprtk::expression<float> x_expr;
-    inline static exprtk::expression<float> y_expr;
+    inline static double x;
+    inline static double y;
+    inline static exprtk::expression<double> x_expr;
+    inline static exprtk::expression<double> y_expr;
     inline static bool valid = true;
 };
